Scrub: Ignore PVs, warn about ignores
Because we can't tell whether they're still in use, storage scrubbing
ignores physical volumes and SSP logical units.

The existing code was explicitly ignoring LUs, but leaving PVs in the
"unknown" category.  This is now fixed.

We are also now emitting a warning log message specifying the name and
type of storage we're ignoring so that the consumer can manually check
and remove it as necessary.

Change-Id: I2577de6bcd03c48888bfe889a0a5e7c9d06329e5