Improve the find vios by port map
Depending on where the port map is generated, the data needed within a
port map for finding the hosting VIOS may change.  This code change
enhances how a VIOS is found to look first at the client WWPNs.  If not
found, then delegates back to the original logic.

This helps with LPM, as the backing physical port from the mapping will
not be on the target server.  But the client WWPNs will be.

Change-Id: I4a1fe10615cb01368d630fe66f7a72536430501a