Fall back to LUARecovery if pg83 not in PV
Current VIOS levels supply the SCSI page 0x83 descriptor in a fresh
QUERY_INVENTORY, but not in a PV ADD Event.  The REST server populates
its cache from Events where possible, so newly-configured physical
volumes will be lacking this attribute.

VIOS is working to remedy this omission; but until that support is
available, a workaround is needed for consumers wishing to make use of
the pg83 identifier.

This change set provides the workaround by invoking QUERY_INVENTORY via
the LUARecovery Job if the pg83 descriptor is not found in the
<PhysicalVolume/>.

Change-Id: Id022c6368dcab9ab34234799fde28f0509767d29