Provide graceful errors on pg83
The pg83 support on the Physical Volume (PV) is a useful wrapper.
However, it relies on the 'parent' object being set.  This will not
occur if the PV is built via the VSCSIMapping object.

This provides an error that is a bit more graceful than an attribute
error that one would normally get.

Change-Id: Ie9e7a80438352a6983cab26ce61fa1e9902d9c4f