Migrate wwpn generation API to REST
The WWPN unique generation API is currently randomly generating WWPNs.
There is a small chance that the WWPNs may overlap within a given
fabric.  While that is very very low, the PowerVM servers to have a set
of globally unique WWPNs that can be used.

This change set migrates the existing API over to use the PowerVM REST
API to generate the guaranteed unique WWPNs.

Change-Id: Ic637d5c2bb2697f5b2fa73ca0fa928dab08aa9e4