Handle when samples just started
This change addresses an issue when the metrics have just been enabled
on the system.  This leads to a timing window where the system does not
have metrics for an instance.  That would return an empty phyp set of
data.

This will return it as an empty set of data.  Future requests will
saturate the data as the REST API starts to obtain the instance data.

Change-Id: I4d8f3fd4f4750599457f24c5ceeef9200a3669e3
Closes-Bug: 1511228