WIP: Provide exact/at least on npiv map
The NPIV mappings may have requirements for 'at least' or 'exactly' a
certain number of VIOSes.  This can be due to a live migration operation
for instance, that requires exactly the same number of VIOSes on the
target.

This change set adds that capability to the derive_npiv_maps function.

Change-Id: I337960cb5d80062446d907f0cec2a7b60c37a438
Closes-Bug: bug/1512229