Subtasks should not lock the against the entry
The entry_transaction decorator locks based on the UUID of the
entry, however there are cases where entry tasks can spawn off
subtasks that will again lock against the entry UUID.  When this
happens it was causing a deadlock case.  The entry_transaction
decorator tries to protect against flooding the entry object
(such as the VIOS) with multiple update requests which would
cause retry operations. In the case where a task spawns off
a subtask that's updating the same entry, we can safely remove
the lock thus avoiding the deadlock case.

Change-Id: If7fe6f2149217b2aba92d74f5f238cf9c16f0028