Type-Agnostic Wrapper Search
The search method implemented in EntryWrapper can take two separate code
paths:

1) If the search parameters are suitable, we can construct a query
through the REST API's search mechanism.  The processing is done on the
server side, and the resulting payload comprises the search results,
which we pass back to the caller.

2) Certain conditions are not supported by the REST API's search
mechanism, including searching for CHILD objects, using extended
attribute groups, and searching on fields not specifically supported by
the REST API for that object.  In such cases, the search method
retrieves the entire feed of objects of the requested type and performs
the search/filtering on the client side.

In code path 1, we're constructing a URI, and the PowerVM API does not
have any type indicators in its search syntax, so everything implicitly
gets 'cast' to string.  For example, if you're looking for a NetBridge
with PVID 1, it doesn't matter whether you search for 1 or '1'.

However, in code path 2, since we're using the wrapper getter methods to
pull out the values to compare against, 1 and '1' were treated
differently.  In fact, in cases where the wrapper getter provides a
non-string type, any search using a string value would be guaranteed to
fail.

This change set fixes the behavior of the search algorithm such that
both code paths behave consistently.  You will now get the same result
for 1 and '1' via either code path.

Change-Id: I16c948511f86e1ea0b0fe5ff459c00af3bbb4847