Change and cleanup timeout configuration values.
There were two job timeout configuration settings. This change is
to get rid of the redundancy.

This change also adds another configuration setting for cancel jobs.
Previously the cancel job was using a default timeout of 1800s which
is quite long. The new value is now 120s.

This change also makes the cancel job's monitoring asynchronous.

Change-Id: Ic0d75238aa003f1c7aa34c14a297bb1b0e3f3b3f