Provide 'remotable' local vnc term
WIP: More unit tests

This change set provides support for a 'remotable' local vnc vterm.  The
standard vnc session within the PowerVM API is only available to
127.0.0.1.  That means that routing of the VNC must be from a consumer
on the localhost (or external is permitted using a password...but not
all consumers can use this).  Localhost consumption is truly rare.

This provides a 'repeater'.  This repeater consumes the localhost VNC
and repeats it externally.  This is useful for very specific scenarios -
ex. you want to send the VNC stream to perhaps only a single IP.

With this repeater the userr can:
 - send the VNC to a specific IP
 - utilization a validation string on connect.  This helps ensure that
   the user is correct, and not a random VNC consumer.

This is not a security measure (as VNC by default is not secure).
Rather it is some additional validation to help ensure that the consumer
of the session is who they say they are.

Change-Id: Iae2f1484bb230a5a8e6fae1899a25393ad3165cd