Fix tox timing issue with cancel_job tests
On rare occassions the test_cancel_job and test_cancel_job_thread
test case assertions will conflict with one another due to
the timings of the child threads spawned in the code.

This change mocks the TestCancelJob's start method in test_cancel_job
instead of the run method so that no child thread is ever
started.

Change-Id: I6af66ce32851ec69a172aceaf5f90b7c697dffa5