More unique session auth file name
Previously the session authentication username was generated using a
seconds-since-the-epoch suffix.  If multiple processes (or threads
within a process) attempted to create a session within the same second,
it was possible to see this error:

OS able to read file
/opt/pvm-rest/data/xapisessionFor_pypowervm_1448939049, but encountered
an access error.

...under certain circumstances:

Thread A: Request session token with username pypowervm_123
Thread B: Request session token with username pypowervm_123
Thread A: Read session file
Thread A: Use session token to perform a REST operation
REST server: (Seeing first use of the token) delete session file
Thread B: Read session file <== error (ENOENT errno=2)

With this change set, we now guarantee uniqueness of the filename by
generating the username with a combination of the thread ID and a
processwide monotonic counter.  The counter ensures that one thread
doesn't collide with itself; and the thread ID ensures it won't collide
with other threads on the system (including those in other processes,
which have their own counters).

Closes-Bug: 1521198
Change-Id: I06a58493d9dca07319614cf6952b99fadaadfebf