FreeIOAdaptersForLinkAggregation in Network xag
When requesting a wrapper property that doesn't exist, pypowervm may
sometimes vivify that property.  In such cases, we have an affordance
for populating the proper attribute data in the XML tag.  In the case of
FreeIOAdaptersForLinkAggregation, we were not populating the 'group'
(extended attribute group) attribute.  Hence any time this field was
vivified and the wrapper was POSTed back to the REST server, one would
see errors like:

REST0001 Failed to unmarshal input payload.

Attribute 'group' must appear on element
'VirtualIOServer:FreeIOAdaptersForLinkAggregation'.

This change set adds the proper 'group' attribute to this field.

Change-Id: I847680256d7ef9ace5ed51409b8b790bcffc40f6