Detach storage from VSCSI mappings
pypowervm.tasks.scsi_mapper.detach_storage is a new method to detach the
storage from the VSCSI mappings within a VIOS wrapper that match
specified criteria.  This is accomplished by replacing matching mappings
with copies lacking the Storage and TargetDevice members.

The caller of this method is responsible for locking and for POSTing the
results back to the REST server.

Change-Id: I71bca29aa6053ee17d80c0d6ec82ece114093eb7