Prevent element poaching when creating new Element
A quirk of lxml (possibly of etree implementations in general) is that
adding an element of one XML hierarchy to a different XML hierarchy
*removes* that element from the first hierarchy.  Pseudocode example:

> len(dom1.children)
3
> dom2.append(dom1.children[0])
> len(dom1.children)
2

pypowervm was being bitten by this quirk: when creating or augmenting a
pypowervm.entities.Element, if a caller passed in one or more member(s)
of some other hierarchy as children, those elements would disappear from
the source hierarchy.

This change set resolves the issue by deep-copying the children before
adding them to the destination Element.

Change-Id: I05a2aaa4599e9f1e2e63ea749eb0ea5b3884e705