Do Not Merge. Patch for local to remote connection
This patch is to enable a VM on the local PVM system to talk
remotely to the PVM VM, while thinking it's talking to the
VM locally.  It patches the code to find an address that is
being used for RMC and use that for the pypowervm host IP.

Requirements:
- install ConfigParser
- enable 'sudo' to work for the current user with requiring
  a password.  Sudo is needed to get the RMC address.

Configuration:
- Put the pypowervm.ini file in the home directory of the
  user if the PVM VM user / password are non-standard.

Unit Tests:
- The unit tests have not been updated to pass.

Do not merge this patch. It's internal only.

Change-Id: I50f92d0e8c2ae8164e7c45d20bbfc6af4204225c