[PATCH] DO NOT MERGE.  Local-to-remote connection
This patch enables the local system to use pypowervm.adapter to talk to
a remote PVM REST server as if it were local.  The patched version of
Session finds a live RMC address and uses that for the pypowervm host IP
address.

Requirements:
- install json
- enable 'sudo' to work for the current user without requiring a
  password.  This is needed to get the RMC address.

Configuration:
- If the PVM REST server credentials are nonstandard, create
  pypowervm.json in the home directory of the user as follows:

{
  "password": "bmVvMTIz",
  "user": "neo"
}

...where the 'password' value is the base64 encoded version of the
cleartext password (echo -n 'cleartext_password' | base64)

Unit tests and pep8 will not pass with this patch in place.

DO NOT MERGE THIS PATCH.  IT IS INTERNAL ONLY.

Change-Id: I50f92d0e8c2ae8164e7c45d20bbfc6af4204225c