Remove the odd validation check scheme
The validation_check, success and fail pattern exposed too much
information to the consumer.  Instead the basis for this is to
essentially have a model where a path is passed in, and if so we know
that the client is using a HTTP request model.

The validation scheme has been dramatically simplified to just require a
vnc path.  This makes it much easier for 'upward' integration.

A callback mechanism was evaluated, but it was later determined that
this scheme (while perhaps slighly less flexible) meets the majority of
clients.  And this also simplifies the client code more than a callback
does.

Change-Id: If62cb554ac350877f0d61fa1f0cd00be933bb72d