Remove circular reference to Session
The APITraits class constructor takes a Session class and stores
it for later use.  The Session is the object that is constructing
the APITraits object, so that causes a circular reference between
the traits object and the session.  That will prevent garbage
collection of the session and hence the session will never be
logged off.  The circular reference can be avoided by using a
weak reference to the session in the APITraits class.

Change-Id: I12ee67e537ca6a24b64c1efd7480769ca7095c27