Remove Session used exclusively for Events
To avoid circular references and allow garbage collection of
the Session object, a second Session object was created
exclusively to handle events.  This is because the EventListner
is hung below a Session. If a single Session is used it causes
a loop:

Session->EventListner->Adapter->Session

However, the reference loop can be broken by using a weak
reference proxy between the Adapter and the Session. Which
is like:

Session->EventListner->Adapter->weakref.ProxyType=>Session

Change-Id: I057c10dffff900f1797c4d7b064172c604904e3e