Allow handlers for raw events from PowerVM.
The existing EventHandler interface doesn't allow a handler access
to the events in their sequential order nor does it allow access
to all of the event data passed from PowerVM. The introduced
RawEventHandler class will fill that gap. When its process()
method is called it will be able to process the events in order
and also have access to all of the event data.

Change-Id: Ic0e1cb4b086d82513c179ba9629229f53efc04cc