Fixes a bad use of the client_addr
The vterm code had a debug statement left in that caused clients to fail
to connect (it was throwing a TypeError).

This change fixes the debug message so that it can print the IP address
of the client properly.  It also fixes the unit tests so that this can't
happen again.  During that process, it also fixed one of the 'bad path'
unit tests.

Change-Id: I7d04563894cefc9b94a337e5ca15c47b6f149582