Unique session username (really, this time)
Commit a2b1b96897c220534d21837f0d57b6e787a247b8 attempted to ensure
uniqueness of the string used to generate a REST session token by
using a thread ID and a processwide monotonic counter.  However, a bug
was retained due to the following "feature" of python syntax:

Reading a variable via 'self' only accesses the class variable if an
instance variable of the same name has not been created.  Writing a
variable via 'self' creates an instance variable, even if the prior read
of that symbol fell through to the class variable.  That is:

class Foo(object):
    x = 1

    def foo(self):
        print(self.x)  # 1
        print(Foo.x)  # 1
        self.x = 2
        print(self.x)  # 2
        print(Foo.x)  # 1

(The author of the original fix was confused because static/class
methods are often - even usually - accessed via 'self' from instance
methods.  The difference being that one does not typically 'write' to
methods.)

In the original change set, the monotonic counter was being incremented
via 'self' - and therefore only being incremented per-instance, not
processwide.

With this change set, we abandon the idea of generating a unique name in
any clever fashion and instead use a randomly-generated UUID string.
The probability of collisions in this case is extremely small.

Change-Id: Id75fa1a1ff5a86163dbbe4a1b71007112db16b8b