Ensure Session clones do not logout session
There are cases where a Session maybe cloned for a short
period of time (like logging). When clones are garbage
collected they were logging out the original Session. This
change keeps track of which object ran the __init__() method
and will not log out the Session if it's a object cloned
without running the __init__().

Also, in this change set, isolate the Session and Adapter
test cases, as they were occasionally failing when run
in parallel.

Change-Id: Ic78e465109495a1c0b79c81f33b70cbda0bb4b88