Point to Point VIF connection
This patch set allows for the creation of a 'point to point' connection
between a VM and a trunk adapter.  This allows for communication between
two VMs, with one being able to bridge out to an external connection via
other means (ex. OVS).

The point to point connection is only guaranteed at creation time.
Another VM could come in later and add itself to the connection.

Change-Id: I4ded52a325e3e565b93e48bfb7496593ba3ecfca