Change ame_enabled setter property to private setter function
The ame_enabled flag is driven by the setting of the expansion
factor. Setting that factor to 0.0 will automatically force the
ame_enabled flag to be False... sending down False for that flag
to the REST layer when the expansion factor != 0.0 will not change
the flag.

Also, add .pydevproject and .project eclipse metadata to .gitignore

Change-Id: Ie3ef8f6ece62f4780f69ccbc73838ed737308014