Remove ame_enabled setter property
The ame_enabled flag is driven by the setting of the expansion
factor in the REST layer. Setting that factor to 0.0 will automatically
force the ame_enabled flag to be False. Sending down False for that
flag to the REST layer when the expansion factor != 0.0 will not
change the flag.

Change-Id: Ie3ef8f6ece62f4780f69ccbc73838ed737308014