Add support for enterprise pools
Add a wrapper for both PowerEnterprisePool and PowerEnterprisePoolMember
entries. Note that these entries are for the most part read-only besides
a few fields (such as mobile proc units and mobile memory).

Provide utility methods for setting the number of mobile proc units and
the amount of mobile memory for a specific pool member.

Change-Id: If4bb76ac63723d81b4397cd5d85cc7c19939add0