Deprecate xag param to EntryWrapper.update()
Deprecate the xag parameter to
pypowervm.wrappers.entry_wrapper.EntryWrapper.update and remove its
influence on that method.

A wrapper must be POSTed using the same extended attribute groups with
which its GET was originally performed; else an etag mismatch (HTTP 412)
is guaranteed to occur.  We should therefore always perform
EntryWrapper.update with the exact URI from the wrapper (the 'href'
property); so the xag parameter was worse than useless.

Furthermore, the xag parameter was being used in such a way that, if it
was omitted, and the original wrapper GET was done with no groups (that
is, xag=[]), we would add ?group=None - thus guaranteeing an etag
mismatch.  Consumers would have needed to work around this problem by
specifying xag=[] to the update method.

Change-Id: Ic696d8942f5567da9bcd28d7c22ef2b218f2ce72