Make EntryWrapper.get work properly for CHILDren
Previously, an attempt to use EntryWrapper.get() to retrieve a CHILD
object would have looked like e.g.

  cna_wrap = LPAR.get(adpt, root_id=lpar_uuid, child_type=CNA.schema_type,
                      child_id=cna_uuid)

The first problem is that you had to ask for a CNA by calling get() on
LPAR.

The second problem is that this internally resulted in:

  resp = adpt.read(... maybe the right stuff to get a CNA ...)
  return LPAR.wrap(resp)

That is, the final wrap was done from LPAR.  This will break (or, worse,
might actually return you an LPAR wrapper with a CNA in it) because resp
contains a ClientNetworkAdapter, not a LogicalPartition.

With this change set, retrieving a CHILD object is now possible via
intuitive syntax, using new 'parent_type' and 'parent_uuid' kwargs.

  cna_wrap = CNA.get(adpt, parent_type=lpar.LPAR, parent_uuid=lpar_uuid)

The parent_type arg may be specified as either the class (as above) or
the string:

  cna_wrap = CNA.get(adpt, parent_type=lpar.LPAR.schema_type,
                     parent_uuid=lpar_uuid)

(The EntryWrapper.search API was also augmented to accept either the
class or the string for parent_type.)

When retrieving an entry rather than a feed, the entry's UUID may be
specified via the new 'uuid' kwarg:

  cna_wrap = CNA.get(adpt, parent_type=lpar.LPAR, parent_uuid=lpar_uuid,
                     uuid=cna_uuid)

...or, for compatibility, via 'child_id':

  cna_wrap = CNA.get(adpt, parent_type=lpar.LPAR, parent_uuid=lpar_uuid,
                     child_id=cna_uuid)

Retrieving ROOT objects is backward compatible - though the new 'uuid'
kwarg is now also available to use in place of 'root_id':

  lpar_wrap = LPAR.get(adpt, uuid=lpar_uuid)

Closes-Bug: 1544146

Change-Id: I38023109ff818a6e883188d24c3bc41a18e4bd8f
(cherry picked from commit 9bf478bb100c211c480aaabfdeed93b3d2b97409)