PartitionNVRAM XAGs changes
Changes to support the LPAR NVRAM Extended Attributes.

This change set also introduces a new decorator,
@pypowervm.wrappers.entry_wrapper.Wrapper.base_pvm_type, for use by
Wrapper subclasses which are not to be instantiated themselves, but
which serve as base classes for actual Wrapper classes.  This decorator
allows such base classes to use e.g. the @xag_property decorator
properly.

Change-Id: I5f5661ff74eadab55bda0ec644a398cb590b9514