search(one_result) and resolve some TODOs
This change set resolves a handful of outstanding TODOs in the code.
The main impact is as follows:

=> Convert manual searching of VSwitch feed to use VSwitch.search
instead.
=> To facilitate the above (and other uses of .search), add the
one_result parameter to the search method.  This is to be used when the
caller expects a single search result, resulting in that single wrapper
or None, as opposed to a (possibly-empty) list.

Change-Id: Idfd09c7bf6124dbde267575cdf9b857504267fca