Fix logging issue on debug messages
Debug messages don't need to have precise variable names when logging.
Often times this leads to a scenario where the message has many %s
elements within it.

A previous change switched many of the debug messages from using the %
to build the message up front to instead use a comma (so that the logger
can determine whether or not to build the message once it has decided
that DEBUG is an active log level).

But in doing so, it did not terminate the message properly.

This change fixes that previous issue, and solves the logging errors
that occurred when DEBUG was turned on for pypowervm.

Change-Id: Iae462f29048c3f88488329517966fe3f882b53bd
Closes-Bug: 1552022