Undo ability to rename SSP LUs
A previous commit publicized LU.name as a @property setter.

It turns out renaming SSP LUs isn't supported.  The PowerVM REST API
interprets the changes in the XML payload as "remove the LU with the old
name; create a new one with the new name".  This obviously leaves the
"renamed" LU with no data.

Change-Id: I5da6fc96746c2e178b02628409205902414c7980