Unpack WrappedFailure from FeedTask
When a FeedTask represents multiple feed elements, execution of that
FeedTask may raise exceptions in more than one of those subflows.  In
such cases, TaskFlow wraps the exceptions in a
taskflow.exceptions.WrappedFailure, which can be iterated to find the
individual exceptions.  Unfortunately, the string representation of a
WrappedFailure only contains the message portion of each exception
therein, and the outer traceback only shows the stack that ran the
FeedTask.  There is thus no way to tell what really triggered the
original exception.

With this change set, FeedTask detects when it gets a WrappedFailure and
unpacks it, collecting the tracebacks from the individual nested
exceptions.  It concatenates these together and raises a new exception,
MultipleExceptionsInFeedTask, containing this collection of tracebacks.
This should allow the consumer to see what actually failed.

Change-Id: I39aa3ad3dc78826ac9ee395c29efba23e53eae5e