Don't attempt to unmarshal */Debug/* URI responses
To provide enablement for */Debug/* URIs, which don't follow the
standard XML formatting of other REST objects, we allow responses from
such URIs to pass through without attempting to unmarshal them (which
would result in an exception).  The consumer of such a response is
expected to use the .body property, understanding its expected format
and parsing it "manually".

Change-Id: I0833e5af8c33b492ac941a3f1b244249aac85d02