Coordinated Image LU Upload
A new API, pypowervm.tasks.cluster_ssp.get_or_upload_image_lu, allows
the consumer to perform an upload of an image LU with a particular name
from multiple independent hosts using the same SSP.  The method operates
on the assumption that the consumer is using a naming convention for the
image LUs such that an LU of the same name will have the exact same spec
(size, content, etc.) no matter where it's created/populated from.  The
creation/upload is coordinated so that only one instance of the LU ends
up on the SSP.

We also introduce a new optional parameter, return_ssp, to
pypowervm.tasks.storage.upload_new_lu.  If True, it causes the method to
return the updated SSP EntryWrapper in addition to the other return
values.  The parameter defaults to False for backward compatibility.

Change-Id: Iea78cdaf96e49085c7282e091badf25898d3e19d