LUEnt: EntryWrapper for LogicalUnit CHILD object
This change set introduces pypowervm.wrappers.storage.LUEnt.  Whereas
the existing LU wrapper is an ElementWrapper used to represent a
<LogicalUnit/> DETAIL object as it exists within a <SharedStoragePool/>,
LUEnt is an EntryWrapper used to represent the LogicalUnit REST type
which is a CHILD of Tier.

LUEnt has all the same methods and properties as LU, plus the additional
ones associated with EntryWrapper, such as uuid.

Change-Id: Ibc10a6296d7b0016b9101dcd525769cb0aaec02c