Create LU on a Tier
This change set allows the existing pypowervm.tasks.storage.crt_lu API
to accept either a SSP or a Tier as its first argument.  The LU is
created using the Tier (the default Tier on the SSP if the first
argument is an SSP).  Since LU is a first-class REST object when
addressed under a Tier, this confers the advantage of allowing us to
PUT/DELETE a Logical Unit without worrying about the SSP's etag.

The API is designed to be *almost* backward compatible with its previous
incarnation.  The difference is that DuplicateLUNameError is no longer
raised under any circumstances.

Change-Id: I507e968b1c262d1c1ff78d4ed517770d1a94a577