WIP: Create LU on a Tier
This change set allows the existing pypowervm.tasks.storage.crt_lu API
to accept either a SSP or a Tier as its first argument.  If the former,
the behavior is unchanged.  If the latter, the LU is created using the
Tier instead.  Since LU is a first-class REST object when addressed
under a Tier, this confers the advantage of allowing us to PUT/DELETE a
Logical Unit without worrying about the SSP's etag.

WIP: UT pending

Change-Id: I507e968b1c262d1c1ff78d4ed517770d1a94a577