Set dev name for CNA on mgmt VM
This change set allows for the device name to be set on the CNA device,
assuming it is assigned to the mgmt VM.  The change actually allows any
CNA to set it at bld time, but the REST API should regulate whether or
not its valid (ex. on the Mgmt Server).

Change-Id: I2fcf9998447de7e68165708f9de3abf8a8516cad