WIP: Tier-based LU removal
pypowervm.tasks.storage.rm_tier_storage is a new API similar in spirit
to rm_ssp_storage; but it deletes LUs based on their existence in a
Tier.

Pro: Because we operate on individual LUs (as opposed to the entire
SSP), we don't need to worry about etags, thus don't need to use
@entry_transaction or @retry.  Thus, in high-traffic scenarios, we don't
have to worry about exhausting our 412 limit.

Con: Since we're not removing a disk LU and its backing image in a
single transaction, it's possible an out-of-band operation will attempt
to link a clone to a backing image we are deleting.  Depending on the
timing, this can have one of two results:

1) If the out-of-band link is performed *before* we remove the image,
the deletion will appear to succeed.  The backing image continues to
exist in a fully-functional but "phantom" (invisible) state until all
clones are removed.

2) If the out-of-band link is attempted *after* the image is removed,
the link will fail.  In this case, the caller should remove the would-be
clone LU, upload her own copy of the image, and link to that.

Con: The network traffic and overall performance of N individual calls
to LogicalUnit DELETE may be slightly worse than that of one
SharedStoragePool POST wherein the same N LU deletions are indicated.

WIP: UT pending

Change-Id: I3dd55ef83ded7e34aeb4a567a95d98337cc048f4