Change retry default to 6 with STEPPED_DELAY
The retry mechanism is used in a lot of the OpenStack code.  It was
found in initial scaling that it was very beneficial in the FeedTask
code that the STEPPED_DELAY and retry count of 6 proved most useful.

With the nova CI testing hitting concurrency in new ways (not from a
single OpenStack but from multiple), we are finding that the stepped
delay function is really the ideal function.

When you have multiple independent processes hitting one API, your
locking mechanisms are limited.  It turns out introducing a slight delay
and increasing the retry count smooth a lot of the issues over.  It
surely does NOT solve the issues, but provides a much smoother test bed.

This update changes the default across the project to use the
STEPPED_DELAY and the retry count of 6 for all retry calls (unless
explicitly overridden).  This is not a full solution.

The usage of the retry decorator within the OpenStack code was checked
and this should be acceptable for that code.  Also the invocations
within pypowervm were validated.  It is not expected that this will
affect client code dramatically, other than to provide a better default
experience in high concurrency environments.

Change-Id: I004d39908f5c8b0a3ac17f29593e815ea43f8b6c