Revert "Change retry default to 6 with STEPPED_DELAY"
This reverts commit 76b1cc23dd2d0625e0e9a6d97e6ea460bc8e50e8.

The retry change wasn't helping where we wanted it to, and it caused compatibility issues in other projects which, while they could be fixed, weren't worth the hassle.

We've decided both the retry count and the change from NO_DELAY to STEPPED_DELAY should be made at the individual  consumers instead.

Change-Id: I23ae62c9e3b09ce0e18af1bc3490e9c0b1998538