Create LU linked clone directly (not through Job)
The REST API used to require LU linked clones to be created by first
creating a blank disk LU; and then invoking the 'LULinkedClone' Job to
link to the backing image LU.

The REST API now supports creating a linked clone in a single step by
having the cloned_from_udid present in the LU(Ent) when it is first
created.

This change set adds to LU(Ent).bld and crt_lu the ability to specify a
backing image LU to be cloned from.  It also deprecates the existing
crt_lu_linked_clone method, which is no longer necessary.

Change-Id: Ia24ab9d836d263a8b9d8e381f4230a0e8256633a