SlotMapBase and RebuildSlotMap
The new classes pypowervm.tasks.slot_map.SlotMapBase and .RebuildSlotMap
facilitate rebuilding an LPAR with the same I/O slot topology.

SlotMapStore:
Non-persisting base class for mapping an LPAR's I/O elements (VFC and
VSCSI storage and CNAs) to the client slots on which they are located.

BuildSlotMap:
The base class that stores the topology and I/O configuration for an
initial VM build.

RebuildSlotMap:
Takes a SlotMapBase.topology and the I/O configuration of a target
system and produces an appropriate allocation of slots on the target
system.

TODO: Expand BuildSlotMap and RebuildSlotMap UT

Change-Id: I20bd814f62675594880b01d1729471d3c2965017