Support next available slot for possible mgmt vif
This change supports the 'next available slot' if the client is
rebuilding a VM with a mgmt CNA.  The mgmt CNA may or may not exist on
the original source server.  If it does not, but the rebuild is
requesting it, then it should be given the next available slot.

The reason that this is part of the Rebuild subclass instead of the main
Build class is because the Build operation will be adding to its slot
map as it goes.  And this can confuse the backing REST API.  By
attaching to the rebuild, it makes sure that this action only occurs
when all slots need to be known.

Change-Id: I5c863967927d028d834dda090cf55fa69c8aff49