Support next available slot for possible mgmt vif
This change supports the 'next available slot' if the client is
rebuilding a VM with a mgmt CNA.  The mgmt CNA may or may not exist on
the original source server.  If it does not, but the rebuild is
requesting it, then it should be given the next available slot.

The reason that this is part of the Rebuild subclass is that the rebuild
assumes that it knows all of the existing slots.  But this is one
scenario where the slot may differ from the source to target.  This
single edge case is better encapsulated in the Rebuild itself.

Change-Id: I5c863967927d028d834dda090cf55fa69c8aff49