Save and restore/derive LPAR's maximum slot number
When restoring an LPAR, we need a way to determine a reasonable maximum
slot number with which to create the LPAR
(pypowervm.wrappers.base_partition.max_virtual_slots).  This change set
allows the max_virtual_slots of the source LPAR to be saved in the
SlotMapStore and retrieved via the BuildSlotMap.  If no value was saved
on the source, BuildSlotMap derives a reasonable value based on adding
10 to the highest slot number saved in the SlotMapStore.

Change-Id: I335b86c4913264877cd50a13b85eaf75d7537aa4
Partial-Bug: 1587942