Increase sleep times on SSP upload
When a single image is being uploaded from multiple sources, there is
orchestration in the pypowervm code to ensure that only one of the
sources actually does the upload and the rest wait for the original
upload to complete.  Once the original upload completes, the others will
clone from that source.

The servers that are waiting for the main to finish uploading poll the
SSP to see if the upload is done.  However, with a sufficient number of
hosts, this can dramatically slow down the upload as it queries the SSP
database quite often.

This change set reduces the polling query to the SSP dramatically.
Instead of polling once every three seconds, it will poll a random value
between 30 and 60 seconds.  This value seems appropriate as waiting for
the other hosts publishing the same image should generally be rare.

Change-Id: I01d466a2a6a7e2da3a4fbe978aef7ba7c20fd3a8
Closes-Bug: 1589308