Add SDN/VLAN Override to migration job
This change set adds the SDN and VLAN Overrides to the migration job.

The SDN override indicates that the invoker is using a technology
outside of Shared Ethernet to drive the live migration and knows what
actions to take. This defaults to false as it is not native to VIOS.

The VLAN override indicates that the invoker does not need the VIOS to
validate that the VLAN is already on the target server.  This is useful
for environments that will provision the vlan on the target as it
performs its migration operations.

Change-Id: Id7ba6bae830142db7d33b980edcba3c00368f23a